+++ ypf_ruta/src/actores/gestor/handlers/modificar_limite_general.rs
@@ -6,6 +6,6 @@ impl Handler<ModificarLimiteGeneral> for Gestor {
     type Result = Result<(), String>;
 
     fn handle(&mut self, msg: ModificarLimiteGeneral, _ctx: &mut Context<Self>) -> Self::Result {
-        self.modificar_limite_general_empresa(msg.id_empresa, msg.nuevo_limite)
+        self.modificar_limite_general_empresa(msg.id_empresa, msg.nuevo_limite as f32)
     }
 }
