+++ ypf_ruta/src/actores/gestor/handlers/modificar_limite_particular.rs
@@ -6,6 +6,6 @@ impl Handler<ModificarLimiteParticular> for Gestor {
     type Result = Result<(), String>;
 
     fn handle(&mut self, msg: ModificarLimiteParticular, _ctx: &mut Context<Self>) -> Self::Result {
-        self.modificar_limite_particular_tarjeta(msg.id_tarjeta, msg.nuevo_limite)
+        self.modificar_limite_particular_tarjeta(msg.id_tarjeta, msg.nuevo_limite as f32)
     }
 }
