+++ ypf_ruta/src/actores/gestor/messages.rs
@@ -1,5 +1,8 @@
+use std::collections::HashMap;
+
 use actix::Message;
-use crate::actores::gestor::structs::{Empresa, Tarjeta, Venta};
+use util::structs::venta::Venta;
+use crate::actores::gestor::structs::{Empresa, Tarjeta};
 
 #[derive(Message)]
 #[rtype(result = "bool")]
@@ -7,19 +10,19 @@ pub struct ValidarVenta(pub Venta);
 
 #[derive(Message)]
 #[rtype(result = "Option<(Empresa, Vec<Tarjeta>)>")]
-pub struct ConsultarEstado(pub u64);
+pub struct ConsultarEstado(pub usize);
 
 #[derive(Message)]
 #[rtype(result = "Result<(), String>")]
 pub struct ModificarLimiteGeneral {
-    pub id_empresa: u64,
+    pub id_empresa: usize,
     pub nuevo_limite: u64,
 }
 
 #[derive(Message)]
 #[rtype(result = "Result<(), String>")]
 pub struct ModificarLimiteParticular {
-    pub id_tarjeta: u64,
+    pub id_tarjeta: usize,
     pub nuevo_limite: u64,
 }
